Visual mode and cleverer change detection
Instead of the undo/redo magic, just remap "c" and get the changed word
before and after. Some magic to get the typeahead is required, but
so far, it seems to work like a charm.

The code probably needs to be refactored a bit to make it easier to add
mappings for "i" and "a".